Bugfix of class cast exception

+++ cpg-library/src/main/java/de/fraunhofer/aisec/cpg/TranslationManager.java
@@ -125,7 +125,7 @@ public class TranslationManager {
             Object sloc =metricMap.get(FileBenchmark.KEY_SOURCE_LOC);
             Object absolutTime = metricMap.get(TranslationManager.class.getSimpleName());
             if(sloc != null && absolutTime != null){
-              metricMap.put("Time_PER_SLOC", 1.0 * ((Integer)absolutTime)/((Integer)sloc));
+              metricMap.put("Time_PER_SLOC", 1.0 * ((Long)absolutTime)/((Long)sloc));
             }
             this.metricMap = metricMap;
             for (Map.Entry<String, Object> entry : metricMap.entrySet()) {
